Move TestCardEncryptor and TestGenericEncryptor to test fixtures
COAND-911
